fix module pattern
