+++ package.json
@@ -4,7 +4,7 @@
   "description": "Returns a list of Governors by state when provided with a list of state abbreviations. Uses the Google Civic Information API.",
   "main": "index.js",
   "scripts": {
-    "fetchGovernors": "node ./index.js",
+    "fetchGovernors": "node ./main.js",
     "test": "jest"
   },
   "repository": {
